fix: Retrieving a single account for detail pages
The issue was that the text's font was in white color.
The information was always being returned.

Used a hacky solution to get the account transaction id and
the such from the code however. Might need to check on a better
method to do so.

The `AccountDetailView` can be removed, but it'll be kept for reference
while the deposit, withdraw, and balance views get created.