feat: Image Slider
Added a slider to the landing page. The slider is a carousel that
displays a series of images. The images are stored in the slider.js
file of the static folder. The slider.js file is imported in the
landing_page.html file. The slider.css file is imported in the
landing_page.html file. The slider.css file is used to style the
slider. The slider.js file is used to create the slider. The slider
is responsive and will adjust to the size of the screen. The slider
is also mobile friendly.